consolidate duplicate code
