+++ src/eduid/common/fastapi/context_request.py
@@ -1,0 +1,56 @@
+from dataclasses import asdict, dataclass
+from typing import Callable, Optional, Union
+
+from fastapi import Request, Response
+from fastapi.routing import APIRoute
+
+
+@dataclass
+class Context:
+    def to_dict(self):
+        return asdict(self)
+
+
+class ContextRequest(Request):
+    def __init__(self, contextClass: type[Context], *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.contextClass = contextClass
+
+    @property
+    def context(self):
+        try:
+            return self.state.context
+        except AttributeError:
+            # Lazy init of self.state.context
+            self.state.context = self.contextClass()
+            return self.context
+
+    @context.setter
+    def context(self, context: Context):
+        self.state.context = context
+
+
+class ContextRequestMixin:
+    @staticmethod
+    def make_context_request(request: Union[Request, ContextRequest], contextClass: type[Context]) -> ContextRequest:
+        if not isinstance(request, ContextRequest):
+            request = ContextRequest(contextClass=contextClass, scope=request.scope, receive=request.receive)
+        return request
+
+
+class ContextRequestRoute(APIRoute, ContextRequestMixin):
+    """
+    Make ContextRequest the default request class
+    """
+
+    # Override in subclass to change the default context class
+    contextClass: type[Context] = Context
+
+    def get_route_handler(self) -> Callable:
+        original_route_handler = super().get_route_handler()
+
+        async def context_route_handler(request: Union[Request, ContextRequest]) -> Response:
+            request = self.make_context_request(request=request, contextClass=self.contextClass)
+            return await original_route_handler(request)
+
+        return context_route_handler
