+++ src/eduid/scimapi/routers/events.py
@@ -3,9 +3,10 @@ from typing import Optional
 
 from fastapi import Response
 
+from eduid.common.fastapi.context_request import ContextRequest
 from eduid.common.models.scim_base import SCIMResourceType
 from eduid.scimapi.api_router import APIRouter
-from eduid.scimapi.context_request import ContextRequest, ContextRequestRoute
+from eduid.scimapi.context_request import ScimApiRoute
 from eduid.scimapi.exceptions import BadRequest, ErrorDetail, NotFound
 from eduid.scimapi.models.event import EventCreateRequest, EventResponse
 from eduid.scimapi.routers.utils.events import db_event_to_response, get_scim_referenced
@@ -16,7 +17,7 @@ __author__ = "lundberg"
 
 
 events_router = APIRouter(
-    route_class=ContextRequestRoute,
+    route_class=ScimApiRoute,
     prefix="/Events",
     responses={
         400: {"description": "Bad request", "model": ErrorDetail},
