+++ src/eduid/scimapi/routers/groups.py
@@ -1,8 +1,9 @@
 from fastapi import Response
 
+from eduid.common.fastapi.context_request import ContextRequest
 from eduid.common.models.scim_base import ListResponse, SCIMResourceType, SearchRequest
 from eduid.scimapi.api_router import APIRouter
-from eduid.scimapi.context_request import ContextRequest, ContextRequestRoute
+from eduid.scimapi.context_request import ScimApiRoute
 from eduid.scimapi.exceptions import BadRequest, ErrorDetail, NotFound
 from eduid.scimapi.models.group import GroupCreateRequest, GroupResponse, GroupUpdateRequest
 from eduid.scimapi.routers.utils.events import add_api_event
@@ -16,7 +17,7 @@ from eduid.scimapi.search import parse_search_filter
 from eduid.userdb.scimapi import EventLevel, EventStatus
 
 groups_router = APIRouter(
-    route_class=ContextRequestRoute,
+    route_class=ScimApiRoute,
     prefix="/Groups",
     responses={
         400: {"description": "Bad request", "model": ErrorDetail},
