+++ src/eduid/scimapi/routers/status.py
@@ -3,15 +3,16 @@ from typing import Mapping
 
 from fastapi import Response
 
+from eduid.common.fastapi.context_request import ContextRequest
 from eduid.scimapi.api_router import APIRouter
-from eduid.scimapi.context_request import ContextRequest, ContextRequestRoute
+from eduid.scimapi.context_request import ScimApiRoute
 from eduid.scimapi.models.status import StatusResponse
 from eduid.scimapi.routers.utils.status import check_mongo, check_neo4j, get_cached_response, set_cached_response
 
 __author__ = "lundberg"
 
 
-status_router = APIRouter(route_class=ContextRequestRoute, prefix="/status")
+status_router = APIRouter(route_class=ScimApiRoute, prefix="/status")
 
 
 @status_router.get("/healthy", response_model=StatusResponse, response_model_exclude_none=True)
