+++ src/eduid/scimapi/routers/utils/events.py
@@ -5,9 +5,9 @@ from uuid import uuid4
 from fastapi import Response
 
 from eduid.common.config.base import DataOwnerName
+from eduid.common.fastapi.context_request import ContextRequest
 from eduid.common.models.scim_base import Meta, SCIMResourceType, SCIMSchema, WeakVersion
 from eduid.common.utils import make_etag, urlappend
-from eduid.scimapi.context_request import ContextRequest
 from eduid.scimapi.exceptions import BadRequest
 from eduid.scimapi.models.event import EventResponse, NutidEventExtensionV1, NutidEventResource
 from eduid.userdb.scimapi import EventLevel, EventStatus, ScimApiEvent, ScimApiEventResource, ScimApiResourceBase
