+++ src/eduid/scimapi/routers/utils/users.py
@@ -6,10 +6,10 @@ from fastapi import Response
 from pymongo.errors import DuplicateKeyError
 
 from eduid.common.config.base import EduidEnvironment
+from eduid.common.fastapi.context_request import ContextRequest
 from eduid.common.models.scim_base import Email, Meta, Name, PhoneNumber, SCIMResourceType, SCIMSchema, SearchRequest
 from eduid.common.models.scim_user import Group, LinkedAccount, NutidUserExtensionV1, Profile, UserResponse
 from eduid.common.utils import make_etag
-from eduid.scimapi.context_request import ContextRequest
 from eduid.scimapi.exceptions import BadRequest
 from eduid.scimapi.routers.utils.events import add_api_event
 from eduid.scimapi.search import SearchFilter
